requirements.txt
# --- For-Bot-Working --------- #
pyrogram
TgCrypto
pyromod
# --- For-Database ------------ #
pymongo
dnspython
# --- For-Web-Response ------- #
aiohttp
python-dotenv
